src/infra/db/mongodb/survey/survey-mongo-repository.spec.ts
import env from '@/main/config/env'
import { SurveyModel } from '@/domain/models/survey'
import { SurveyMongoRepository } from './survey-mongo-repository'
import { MongoHelper } from '../helpers/mongo'
import { Collection } from 'mongodb'
import MockDate from 'mockdate'

let surveyCollection: Collection

describe('Survey Mongo Repository', () => {

  beforeAll(async () => {
    await MongoHelper.connect(env.mongoTest)
  })

  beforeEach(async () => {
    surveyCollection = await MongoHelper.getCollection('surveys')
    await surveyCollection.deleteMany({})
    MockDate.set(new Date())
  })

  afterAll(async () => {
    await MongoHelper.disconnect()
    MockDate.reset()
  })

  const surveyModelData = {
    question: 'any_question',
    answers: [{
      image: 'any_image',
      answer: 'any_answer'
    }],
    date: new Date()
  }

  const makeSut = (): SurveyMongoRepository => {
    return new SurveyMongoRepository()
  }

  describe('add()', () => {
    test('Should add a survey on success', async () => {
      const sut: SurveyMongoRepository = makeSut()
      await sut.add(surveyModelData)
      const survey = await surveyCollection.findOne({ question: surveyModelData.question })
      expect(survey).toBeTruthy()
    })
  })

  describe('loadAll()', () => {
    test('Should load all survey on success', async () => {
      await surveyCollection.insertMany([surveyModelData, {
        question: 'other_question',
        answers: [{
          image: 'other_image',
          answer: 'other_answer'
        }],
        date: new Date()
      }])
      const sut: SurveyMongoRepository = makeSut()
      const surveys: SurveyModel[] = await sut.loadAll()
      expect(surveys.length).toBe(2)
      expect(surveys[0].question).toBe('any_question')
      expect(surveys[1].question).toBe('other_question')
    })

    test('Should load empty list', async () => {
      const sut: SurveyMongoRepository = makeSut()
      const surveys: SurveyModel[] = await sut.loadAll()
      expect(surveys.length).toBe(0)
    })
  })

})